fix docker CI build error frontend
